[BRK] Continued breaking changes
Continuing the total overhaul of the API to be a bit more streamlined, less
clunky. Still not technically working.
