+++ peakdet/physio.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 import numpy as np
-from peakdet import utils, editor
+from peakdet import utils
 
 
 class Physio():
@@ -19,7 +19,7 @@ class Physio():
     def __init__(self, data, fs=None):
         data = np.asarray(data).squeeze()
         if data.ndim > 1:
-            raise ValueError('Provided data dimensionality {} exceeds 1'
+            raise ValueError('Provided data dimensionality {} > expected 1'
                              .format(data.ndim))
         self._fs = np.float64(fs)
         self._data = data
@@ -53,6 +53,9 @@ class Physio():
         """ Sampling rate of data (Hz) """
         return self._fs
 
+    def plot(self):
+        raise NotImplementedError
+
 
 class PeakFinder(Physio):
     """
@@ -64,16 +67,12 @@ class PeakFinder(Physio):
         Input data filename or array
     fs : float
         Sampling rate (Hz) of ``data``
-    col : int, optional
-        Column of data in input file, if ``data`` is str. Default: 0
-    header : bool, optional
-        Whether ``data`` has a header, if ``data`` is str. Default: False
     """
 
-    def __init__(self, data, fs, col=0, header=False):
-        super().__init__(data, fs, col, header)
+    def __init__(self, data, fs=None):
+        super().__init__(data, fs)
         self._peakinds, self._troughinds = [], []
-        self._rejected = np.empty(0, dtype='int')
+        self._rejected = np.empty(0, dtype=int)
 
     @property
     def _masked(self):
@@ -87,16 +86,12 @@ class PeakFinder(Physio):
         if len(self.peakinds):
             diff = (self._masked[:-1] + self._masked[1:]) / (2 * self.fs)
             return diff.compressed()
-        else:
-            return
 
     @property
     def rrint(self):
         """ Length of R-R intervals (in seconds) """
         if len(self.peakinds):
             return (np.diff(self._masked) / self.fs).compressed()
-        else:
-            return
 
     @property
     def peakinds(self):
@@ -108,51 +103,24 @@ class PeakFinder(Physio):
         """ Indices of detected troughs in data """
         return self._troughinds
 
-    @property
-    def _peaksig(self):
-        if self.rrtime is None:
-            return
-        else:
-            return utils.gen_temp(self.data, self.peakinds)
-
-    @property
-    def _template(self):
-        if self.rrtime is None:
-            return
-        else:
-            return utils.corr_template(self._peaksig)
-
-    def reset(self, hard=False):
-        """
-        Removes filtering and peak-finding from data
-
-        Parameters
-        ----------
-        hard : bool (False)
-            also remove interpolation from data
-        """
-
-        self._peakinds, self._troughinds = [], []
-        super(PeakFinder, self).reset(hard=hard)
-
     def get_peaks(self, thresh=0.2, dist=None):
         """
         Detects peaks in data
 
         Parameters
         ----------
-        thresh : float [0,1]
-            determines relative height of data to be considered a peak
+        thresh : float [0,1], optional
+            Determines relative height of data to be considered a peak.
+            Default: 0.2
+        dist : int, optional
         """
 
-        if dist is None:
-            dist = int(self.fs / 4)
-        locs = utils.peakfinder(self.data,
-                                dist=dist,
+        locs = utils.find_peaks(self.data,
+                                dist=self.fs // 4 if dist is None else dist,
                                 thresh=thresh)
-        self._peakinds = utils.peakfinder(self.data,
-                                          dist=round(np.diff(locs).mean()) / 2,
-                                          thresh=thresh).astype('int')
+        self._peakinds = utils.find_peaks(self.data,
+                                          dist=np.diff(locs).mean() // 2,
+                                          thresh=thresh).astype(int)
 
         self.get_troughs(thresh=thresh, dist=dist)
 
@@ -166,54 +134,15 @@ class PeakFinder(Physio):
             determines relative height of data to be considered a trough
         """
 
-        if dist is None:
-            dist = int(self.fs / 4)
         if self.rrtime is None:
             self.get_peaks(thresh=thresh, dist=dist)
 
-        locs = utils.troughfinder(self.data,
-                                  dist=dist,
+        locs = utils.find_troughs(self.data,
+                                  dist=self.fs // 4 if dist is None else dist,
                                   thresh=thresh)
-        troughinds = utils.troughfinder(self.data,
-                                        dist=round(np.diff(locs).mean()) / 2,
+        troughinds = utils.find_troughs(self.data,
+                                        dist=np.diff(locs).mean() // 2,
                                         thresh=thresh)
         self._troughinds = utils.check_troughs(self.data,
-                                               troughinds,
-                                               self.peakinds)
-
-    def plot(self, _debug=False):
-        """ Generates plot of data with detected peaks/troughs (if any) """
-
-        import matplotlib.pyplot as plt
-
-        fig, ax = plt.subplots(1)
-        ax.plot(self.time, self.data, 'b')
-
-        if len(self.peakinds):
-            ax.plot(self.time[self.peakinds],
-                    self.data[self.peakinds], '.r')
-        if len(self.troughinds):
-            ax.plot(self.time[self.troughinds],
-                    self.data[self.troughinds], '.g')
-
-        if not _debug:
-            plt.show()
-
-        return fig
-
-    def edit_peaks(self, _xlim=None):
-        """
-        Opens up peakdet.editor.PeakEditor window
-
-        This is for interactive editing of detected peaks (i.e., is to be used
-        to remove components of the data stream contaminated by artifact). To
-        use, simply drag the cursor over parts of the data to remove them from
-        consideration.
-
-        Accepted inputs
-        ---------------
-        <ctrl-z> : undo last edit
-        <ctrl-q> : stop interactive peak editing
-        """
-
-        editor.PeakEditor(self, _xlim=_xlim)
+                                               self.peakinds,
+                                               troughinds)
