+++ peakdet/tests/__init__.py
@@ -1,0 +1,3 @@
+from peakdet.tests.utils import get_test_data_path
+
+__all__ = ['get_test_data_path']
