+++ peakdet/tests/utils.py
@@ -1,0 +1,9 @@
+"""
+Utilities for testing
+"""
+
+from os.path import join, dirname
+
+
+def get_test_data_path():
+    return join(dirname(__file__), 'data')
